Normalize doc indentation
We used to trim every doc line. This is a problem for code blocks where
the indentation is relevant (e.g. in the amp-list sample). We now
normalize docs based on the indentation of the comment tags, e.g. for:

__<--
__  text
__text
__-->

all whitespaces marked by '_' will be removed resulting in

"
  text
text
"

Fixes #68