add amp-analytics sample
This sample gives an interactive demonstration of howamp-analytics work. We use a realtime dashboard to show which events are fired.

Each user has a separate dashboard which is identified by the AMP CID. The CID is read from the cookie which is not available on the AMP Cache. If the sample is loaded via the AMP Cache, the dashboard will show all user events.

The sample does not set or store any cookies.