Fixed Social Share example (data-param)
I also cleaned it up on several places, there is no need for a width=60 if the default is 60px for example (maybe at the time this example was created there was no default?). Displaying less attributes makes the examples easier to read, especially if they aren't needed for a particular example.

The Basic Usage is a lot simpler this way, ready to go.

The Configuration paragraph now clearly tells the viewer that "text" is the same as the attribute, and I've fixed the data-attributes, because they're (now) supposed to be "data-param-*".

I've started cleaning up the "Customized Views", just stripping away the configuration that isn't needed in this example for now (the example looked quite scary with about 10-15 tags and attributes.). The example bugs out on my screen, so I'm thinking this example is deprecated.