Fixed Social Share example (data-param) (#494)
* Fixed Social Share example (data-param)

I also cleaned it up on several places, there is no need for a width=60 if the default is 60px for example (maybe at the time this example was created there was no default?). Displaying less attributes makes the examples easier to read, especially if they aren't needed for a particular example.

The Basic Usage is a lot simpler this way, ready to go.

The Configuration paragraph now clearly tells the viewer that "text" is the same as the attribute, and I've fixed the data-attributes, because they're (now) supposed to be "data-param-*".

I've started cleaning up the "Customized Views", just stripping away the configuration that isn't needed in this example for now (the example looked quite scary with about 10-15 tags and attributes.). The example bugs out on my screen, so I'm thinking this example is deprecated.

* amp-social-share, Fixed "Customized Views"

The current example doesn't work (anymore), so a new example was required. Because the default styling is done with background images, it's more consistent to overwrite the default CSS than it is to use different element's to hack a similar effect together.

* Fixed SocialShare example updates

Wanted to show the amp-custom values next to the paragraph, but forgot that the validator requires style to only be in the <head>.

* added requested changes

Cleaned up the language a bit, fixed a layout issue and removed the
extra example for Custom Style.

* Removed width&height from Custom Style example