Added an example highlighting the best practices for supporting internationalization of a simple site structure with canonical and AMP documents using hreflang.
Added an explanation for this example as a new documentation page.

Added an explanation for a second example which has yet to be completed.

This second example will highlight a more complex structure including a
canonical, mobile alternate and AMP documents.

Also added two images to explain these examples at a high level.