Improve load performance
- Adds link rel="preload" to the `validator.js` Javascript.
- Adds async to the `validator.js` script.

Since the JS file is a bit large and the script tag is in the end
of the file this will make the request for the validator to start
as soon as the head of the file is received.

It seems Polymer makes the UI visible after domInteractive is
fired. Making validator.js async allows it to fire before it's
downloaded and executed.