Update amp-iframe example
Updated amp-iframe example code by adding `title` attribute to all six `amp-iframe` samples.

[W3C H64: Using the `title` attribute of the `frame` and `iframe` elements](https://www.w3.org/TR/WCAG20-TECHS/H64.html): 
"`title` attribute of the `frame` or `iframe` element to describe the contents of each frame. This provides a label for the frame so users can determine which frame to enter and explore in detail. It does not label the individual page (frame) or inline frame (iframe) in the frameset."

[WCAG 2.4.1 Bypass Blocks](https://www.w3.org/TR/2016/NOTE-UNDERSTANDING-WCAG20-20161007/navigation-mechanisms-skip.html): A mechanism is available to bypass blocks of content that are repeated on multiple Web pages. (Level A)