Updated dimensions to account for recent devices
The viewer should now account for some of the latest devices, at the moment this just uses Apple and Google phones. I didn't add the iPhone XR or Pixel 3 devices, but they could probably be added as well quite easily. To create this I used Google Chrome Dev Tools as inspiration, and verified a few sizes with https://www.paintcodeapp.com/news/ultimate-guide-to-iphone-resolutions.